Renamed new source type from 'sourcefile' to 'file'. Make file extraction optional in fallback case. Added test case suggested in pull request.

+++ eos/fallback.py
@@ -19,8 +19,8 @@ import eos.constants
 import eos.util
 
 
-def download_and_extract_from_fallback_url(fallback_url, filename, relative_src_dir, download_dir, extract_dir,
-                                           sha1_hash_expected=None):
+def download_from_fallback_url(fallback_url, filename, relative_src_dir, download_dir, extract_dir,
+                               sha1_hash_expected=None):
     p = urlparse(fallback_url)
     new_path = p[2] + "/" + relative_src_dir + "/" + filename
     fallback_download_url = urlunparse([p[0], p[1], new_path, p[3], p[4], p[5]])
@@ -31,12 +31,13 @@ def download_and_extract_from_fallback_url(fallback_url, filename, relative_src_
         eos.log_error("downloading of file from fallback URL " + fallback_download_url + " failed")
         return False
 
-    if os.path.exists(extract_dir):
-        shutil.rmtree(extract_dir)
+    if extract_dir is not None:
+        if os.path.exists(extract_dir):
+            shutil.rmtree(extract_dir)
 
-    if not eos.archive.extract_file(download_filename, extract_dir):
-        eos.log_error("extraction of file " + download_filename + " from fallback URL " + fallback_download_url
-                      + " failed")
-        return False
+        if not eos.archive.extract_file(download_filename, extract_dir):
+            eos.log_error("extraction of file " + download_filename + " from fallback URL " + fallback_download_url
+                          + " failed")
+            return False
 
     return True
